ajustement de l'organisation du code pour le répartir dans des fichiers distincts
