+++ utils.h
@@ -1,0 +1,9 @@
+
+#ifndef UTILS_H
+#define UTILS_H
+
+#include "tetris.h"
+
+char valueinarray(int val, int *arr, size_t n);
+
+#endif // UTILS_H
